altura fix

+++ esquema.sql
@@ -79,7 +79,7 @@ create table prateleira(
     nro smallint,
     num_serie   float,
     fabricante  varchar(255),
-    altura  decimal(2,2),
+    altura  smallint,
     nome_cat    varchar(255),
     primary key(nro, num_serie,fabricante),
     foreign key(num_serie, fabricante) references IVM(num_serie, fabricante),
